T7 | new tests clases for more organization
