+++ src/test/java/modelo/BooleanQuestionTest.java
@@ -1,0 +1,46 @@
+package modelo;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class BooleanQuestionTest {
+
+    @Test
+    public void testBooleanQuestionIncreasePlayerPointsWhenOptionIsCorrect() {
+        // Given
+        List<Option> options = Arrays.asList(new Option("si", new CorrectOptionScorer()));
+        QuestionScorer scorer = new BooleanScorer();
+        Question question = new BooleanQuestion("vamos a aprobar algoritmos 3?", options, scorer);
+
+        Player player = new Player();
+        Integer expectedPlayerPoints = 1;
+
+        // When
+        question.score(player);
+
+        // Then
+        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
+    }
+
+    @Test
+    public void testBooleanQuestionDontIncreasePlayerPointsWhenOptionIsIncorrect() {
+        // Given
+        List<Option> options = Arrays.asList(new Option("no", new IncorrectOptionScorer()));
+        QuestionScorer scorer = new BooleanScorer();
+        Question question = new BooleanQuestion("vamos a aprobar algoritmos 3?", options, scorer);
+
+        Player player = new Player();
+        player.setPoints(5);
+        Integer expectedPlayerPoints = 5;
+
+        // When
+        question.score(player);
+
+        // Then
+        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
+    }
+
+}
