+++ src/test/java/modelo/GroupChoiceTest.java
@@ -1,0 +1,4 @@
+package modelo;
+
+public class GroupChoiceTest {
+}
