+++ src/test/java/modelo/MultipleChoiceTest.java
@@ -1,0 +1,34 @@
+package modelo;
+
+import exceptions.InvalidSizeException;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class MultipleChoiceTest {
+
+    @Test
+    public void testMultipleChoiceQuestionIncreasePlayerPointsWhenAllOptionsAreCorrect() throws InvalidSizeException {
+        // Given
+        List<Option> options = Arrays.asList(
+                new Option("2 + 2", new CorrectOptionScorer()),
+                new Option("2 * 2", new CorrectOptionScorer()),
+                new Option("1 + 3", new CorrectOptionScorer()),
+                new Option("2^2", new CorrectOptionScorer()));
+
+        QuestionScorer scorer = new MultipleChoiceScorer();
+        Question question = new MultipleChoiceQuestion("elegir las opciones que dan como resultado igual a 4", options, scorer);
+
+        Player player = new Player();
+        Integer expectedPlayerPoints = 1;
+
+        // When
+        question.score(player);
+
+        // Then
+        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
+    }
+
+}
