--- src/test/java/modelo/QuestionTest.java
@@ -1,134 +1,0 @@
-package modelo;
-
-import exceptions.InvalidSizeException;
-import org.junit.Test;
-import org.junit.Assert;
-
-import java.util.Arrays;
-import java.util.List;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
-public class QuestionTest {
-    @Test
-    public void testBooleanQuestionIncreasePlayerPointsWhenOptionIsCorrect() {
-        // Given
-        List<Option> options = Arrays.asList(new Option("si", new CorrectOptionScorer()));
-        QuestionScorer scorer = new BooleanScorer();
-        Question question = new BooleanQuestion("vamos a aprobar algoritmos 3?", options, scorer);
-
-        Player player = new Player();
-        Integer expectedPlayerPoints = 1;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-
-    @Test
-    public void testBooleanQuestionDontIncreasePlayerPointsWhenOptionIsIncorrect() {
-        // Given
-        List<Option> options = Arrays.asList(new Option("no", new IncorrectOptionScorer()));
-        QuestionScorer scorer = new BooleanScorer();
-        Question question = new BooleanQuestion("vamos a aprobar algoritmos 3?", options, scorer);
-
-        Player player = new Player();
-        player.setPoints(5);
-        Integer expectedPlayerPoints = 5;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-
-    @Test
-    public void testBooleanQuestionWithPenaltyDecreasePlayerPointsWhenOptionIsIncorrect() {
-        // Given
-        List<Option> options = Arrays.asList(new Option("si", new IncorrectOptionScorer()));
-        QuestionScorer scorer = new BooleanWithPenaltyScorer();
-        Question question = new BooleanQuestion("vamos a aprobar algoritmos 3?", options, scorer);
-
-        Player player = new Player();
-        player.setPoints(5);
-        Integer expectedPlayerPoints = 4;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-
-    @Test
-    public void testMultipleChoiceQuestionIncreasePlayerPointsWhenAllOptionsAreCorrect() throws InvalidSizeException {
-        // Given
-        List<Option> options = Arrays.asList(
-                new Option("2 + 2", new CorrectOptionScorer()),
-                new Option("2 * 2", new CorrectOptionScorer()),
-                new Option("1 + 3", new CorrectOptionScorer()),
-                new Option("2^2", new CorrectOptionScorer()));
-
-        QuestionScorer scorer = new MultipleChoiceScorer();
-        Question question = new MultipleChoiceQuestion("elegir las opciones que dan como resultado igual a 4", options, scorer);
-
-        Player player = new Player();
-        Integer expectedPlayerPoints = 1;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-
-    @Test
-    public void testMultipleChoiceQuestionWithPartialScoreIncreasePlayerPointsForeachOptionThatIsCorrect() throws InvalidSizeException {
-        // Given
-        List<Option> options = Arrays.asList(
-                new Option("2 + 2", new CorrectOptionScorer()),
-                new Option("2 * 2", new CorrectOptionScorer()),
-                new Option("1 + 3", new CorrectOptionScorer()),
-                new Option("2^2", new CorrectOptionScorer()));
-
-        QuestionScorer scorer = new MultipleChoiceWithPartialScorer();
-        Question question = new MultipleChoiceQuestion("elegir las opciones que dan como resultado igual a 4", options, scorer);
-
-        Player player = new Player();
-        Integer expectedPlayerPoints = 4;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-
-    @Test
-    public void testMultipleChoiceQuestionWithPenaltyIncreasePlayerPointsWhenOptionIsCorrectAndDecreaseWhenOptionIsIncorrect() throws InvalidSizeException {
-        // Given
-        List<Option> options = Arrays.asList(
-                new Option("2 + 2", new CorrectOptionScorer()),
-                new Option("3 / 1", new IncorrectOptionScorer()),
-                new Option("5 / 1", new IncorrectOptionScorer()),
-                new Option("1 / 1", new IncorrectOptionScorer()),
-                new Option("2^2", new CorrectOptionScorer()));
-
-        QuestionScorer scorer = new MultipleChoiceWithPenaltyScorer();
-        Question question = new MultipleChoiceQuestion("elegir las opciones que dan como resultado igual a 4", options, scorer);
-
-        Player player = new Player();
-        player.setPoints(7);
-        Integer expectedPlayerPoints = 6;
-
-        // When
-        question.score(player);
-
-        // Then
-        Assert.assertEquals(player.getPoints(), expectedPlayerPoints);
-    }
-}
